+++ src/hooks/useLocalVideoToggle/useLocalVideoToggle.test.tsx
@@ -1,9 +1,9 @@
-import { renderHook } from '@testing-library/react-hooks';
+import { act, renderHook } from '@testing-library/react-hooks';
 import useLocalVideoToggle from './useLocalVideoToggle';
 import useVideoContext from '../useVideoContext/useVideoContext';
-import { useAppState } from '../../state';
 import { EventEmitter } from 'events';
 import { LocalParticipant } from 'twilio-video';
+import { useAppState } from '../../state';
 
 import { useParticipantMeta } from '../../withHooks/useParticipantMeta/useParticipantMeta';
 jest.mock('../../withHooks/useParticipantMeta/useParticipantMeta');
@@ -18,14 +18,21 @@ jest.mock('../../state');
 const mockUseAppState = useAppState as jest.Mock<any>;
 mockUseAppState.mockImplementation(() => ({ setError: () => ({}) }));
 
+function getMockTrack(name: string, deviceId?: string) {
+  return {
+    name,
+    mediaStreamTrack: {
+      getSettings: () => ({
+        deviceId,
+      }),
+    },
+  };
+}
+
 describe('the useLocalVideoToggle hook', () => {
   it('should return true when a localVideoTrack exists', () => {
     mockUseVideoContext.mockImplementation(() => ({
-      localTracks: [
-        {
-          name: 'camera',
-        },
-      ],
+      localTracks: [getMockTrack('camera-123456')],
       room: { localParticipant: {} },
     }));
 
@@ -35,11 +42,7 @@ describe('the useLocalVideoToggle hook', () => {
 
   it('should return false when a localVideoTrack does not exist', () => {
     mockUseVideoContext.mockImplementation(() => ({
-      localTracks: [
-        {
-          name: 'microphone',
-        },
-      ],
+      localTracks: [getMockTrack('microphone')],
       room: { localParticipant: {} },
     }));
 
@@ -47,26 +50,24 @@ describe('the useLocalVideoToggle hook', () => {
     expect(result.current).toEqual([false, expect.any(Function)]);
   });
 
-  describe('toggleAudioEnabled function', () => {
-    it('should call track.stop when a localVideoTrack exists', () => {
-      const mockLocalTrack = {
-        name: 'camera',
-        stop: jest.fn(),
-      };
+  describe('toggleVideoEnabled function', () => {
+    it('should call removeLocalVideoTrack when a localVideoTrack exists', () => {
+      const mockRemoveLocalVideoTrack = jest.fn();
 
       mockUseVideoContext.mockImplementation(() => ({
-        localTracks: [mockLocalTrack],
+        localTracks: [getMockTrack('camera')],
         room: { localParticipant: null },
+        removeLocalVideoTrack: mockRemoveLocalVideoTrack,
       }));
 
       const { result } = renderHook(useLocalVideoToggle);
       result.current[1]();
-      expect(mockLocalTrack.stop).toHaveBeenCalled();
+      expect(mockRemoveLocalVideoTrack).toHaveBeenCalled();
     });
 
     it('should call localParticipant.unpublishTrack when a localVideoTrack and localParticipant exists', () => {
       const mockLocalTrack = {
-        name: 'camera',
+        ...getMockTrack('camera-123456'),
         stop: jest.fn(),
       };
 
@@ -76,6 +77,7 @@ describe('the useLocalVideoToggle hook', () => {
       mockUseVideoContext.mockImplementation(() => ({
         localTracks: [mockLocalTrack],
         room: { localParticipant: mockLocalParticipant },
+        removeLocalVideoTrack: () => {},
       }));
 
       const { result } = renderHook(useLocalVideoToggle);
@@ -83,7 +85,7 @@ describe('the useLocalVideoToggle hook', () => {
       expect(mockLocalParticipant.unpublishTrack).toHaveBeenCalledWith(mockLocalTrack);
     });
 
-    it('should call getLocalVideoTrack when a localVideoTrack does not exist', () => {
+    it('should call getLocalVideoTrack when a localVideoTrack does not exist', async () => {
       const mockGetLocalVideoTrack = jest.fn(() => Promise.resolve());
       mockUseVideoContext.mockImplementation(() => ({
         localTracks: [],
@@ -91,12 +93,15 @@ describe('the useLocalVideoToggle hook', () => {
         room: {},
       }));
 
-      const { result } = renderHook(useLocalVideoToggle);
-      result.current[1]();
+      const { result, waitForNextUpdate } = renderHook(useLocalVideoToggle);
+      act(() => {
+        result.current[1]();
+      });
+      await waitForNextUpdate();
       expect(mockGetLocalVideoTrack).toHaveBeenCalled();
     });
 
-    it('should call mockLocalParticipant.publishTrack when a localVideoTrack does not exist and localParticipant does exist', done => {
+    it('should call mockLocalParticipant.publishTrack when a localVideoTrack does not exist and localParticipant does exist', async done => {
       const mockGetLocalVideoTrack = jest.fn(() => Promise.resolve('mockTrack'));
 
       const mockLocalParticipant = new EventEmitter() as LocalParticipant;
@@ -108,12 +113,90 @@ describe('the useLocalVideoToggle hook', () => {
         room: { localParticipant: mockLocalParticipant },
       }));
 
-      const { result } = renderHook(useLocalVideoToggle);
-      result.current[1]();
+      const { result, waitForNextUpdate } = renderHook(useLocalVideoToggle);
+      act(() => {
+        result.current[1]();
+      });
+      await waitForNextUpdate();
       setImmediate(() => {
-        expect(mockLocalParticipant.publishTrack).toHaveBeenCalledWith('mockTrack');
+        expect(mockLocalParticipant.publishTrack).toHaveBeenCalledWith('mockTrack', { priority: 'low' });
         done();
       });
+    });
+
+    it('should not call mockLocalParticipant.publishTrack when isPublishing is true', async () => {
+      const mockGetLocalVideoTrack = jest.fn(() => Promise.resolve('mockTrack'));
+
+      const mockLocalParticipant = new EventEmitter() as LocalParticipant;
+      mockLocalParticipant.publishTrack = jest.fn();
+
+      mockUseVideoContext.mockImplementation(() => ({
+        localTracks: [],
+        getLocalVideoTrack: mockGetLocalVideoTrack,
+        room: { localParticipant: mockLocalParticipant },
+      }));
+
+      const { result, waitForNextUpdate } = renderHook(useLocalVideoToggle);
+      act(() => {
+        result.current[1]();
+      });
+      result.current[1](); // Should be ignored because isPublishing is true
+      expect(mockGetLocalVideoTrack).toHaveBeenCalledTimes(1);
+      await waitForNextUpdate();
+    });
+
+    it('should call onError when publishTrack throws an error', async () => {
+      const mockGetLocalVideoTrack = jest.fn(() => Promise.resolve('mockTrack'));
+      const mockOnError = jest.fn();
+
+      const mockLocalParticipant = new EventEmitter() as LocalParticipant;
+      mockLocalParticipant.publishTrack = jest.fn(() => Promise.reject('mockError'));
+
+      mockUseVideoContext.mockImplementation(() => ({
+        localTracks: [],
+        getLocalVideoTrack: mockGetLocalVideoTrack,
+        room: { localParticipant: mockLocalParticipant },
+        onError: mockOnError,
+      }));
+
+      const { result, waitForNextUpdate } = renderHook(useLocalVideoToggle);
+      act(() => {
+        result.current[1]();
+      });
+      await waitForNextUpdate();
+      expect(mockOnError).toHaveBeenCalledWith('mockError');
+    });
+
+    it('should call getLocalVideoTrack with the deviceId of the previously active track', async () => {
+      const mockGetLocalVideoTrack = jest.fn(() => Promise.resolve('mockTrack'));
+
+      mockUseVideoContext.mockImplementation(() => ({
+        localTracks: [getMockTrack('camera', 'testDeviceID')],
+        room: { localParticipant: null },
+        removeLocalVideoTrack: jest.fn(),
+        getLocalVideoTrack: mockGetLocalVideoTrack,
+      }));
+
+      const { result, rerender, waitForNextUpdate } = renderHook(useLocalVideoToggle);
+
+      // Remove existing track
+      result.current[1]();
+
+      mockUseVideoContext.mockImplementation(() => ({
+        localTracks: [],
+        room: { localParticipant: null },
+        removeLocalVideoTrack: jest.fn(),
+        getLocalVideoTrack: mockGetLocalVideoTrack,
+      }));
+      rerender();
+
+      await act(async () => {
+        // Get new video track
+        result.current[1]();
+        await waitForNextUpdate();
+      });
+
+      expect(mockGetLocalVideoTrack).toHaveBeenCalledWith({ deviceId: { exact: 'testDeviceID' } });
     });
   });
 });