+++ src/hooks/useParticipants/useParticipants.test.tsx
@@ -1,5 +1,5 @@
 /**
- * #ANGLES_EDIT
+ * #WITH_EDIT
  *
  * Removed tests for dominant speaker shuffling, since that got removed from `useParticipants`
  */