+++ src/types.ts
@@ -36,3 +36,5 @@ export type Callback = (...args: any[]) => void;
 export type ErrorCallback = (error: TwilioError) => void;
 
 export type IVideoTrack = LocalVideoTrack | RemoteVideoTrack;
+
+export type LocationTuple = [number, number];