+++ src/withComponents/WidgetProvider/widgetReducer.ts
@@ -1,5 +1,6 @@
 import { Action } from '../RoomState/RoomStateProvider';
 import { WidgetTypes } from './widgetTypes';
+import { LocationTuple } from '../../types';
 
 // `IWidgetProperties` defines the properties that a widget must have.
 export interface IWidgetProperties {
@@ -7,6 +8,7 @@ export interface IWidgetProperties {
   id: string;
   participantSid: string;
   data: any;
+  location?: LocationTuple;
 }
 // `IWidgetState` is the state maintained by this context provider representing the widgets present in the room.
 export interface IWidgetState {
@@ -16,6 +18,9 @@ export interface IWidgetState {
 enum Actions {
   WidgetAdd = 'WIDGET_ADD',
   WidgetRemove = 'WIDGET_REMOVE',
+  WidgetLocationUpdate = 'WIDGET_LOCATION_UPDATE',
+  WidgetUpdateData = 'WIDGET_DATA_UPDATE',
+  WidgetDataFieldUpdate = 'WIDGET_DATA_FIELD_UPDATE',
 }
 
 export default function reducer(state: IWidgetState = {}, action: Action) {
@@ -35,6 +40,47 @@ export default function reducer(state: IWidgetState = {}, action: Action) {
 
       return newWidgets;
     }
+
+    case Actions.WidgetLocationUpdate: {
+      const newWidgets = {
+        ...state,
+        [payload.widgetId]: {
+          ...state[payload.widgetId],
+          location: payload.location,
+        },
+      };
+
+      return newWidgets;
+    }
+
+    case Actions.WidgetUpdateData: {
+      // this action just re-writes the entire widgetData
+      const updatedWidgets = {
+        ...state,
+        [payload.widgetId]: {
+          ...state[payload.widgetId],
+          data: payload.data,
+        },
+      };
+
+      return updatedWidgets;
+    }
+
+    case Actions.WidgetDataFieldUpdate: {
+      const { widgetId, field, value } = payload;
+      const updatedWidgets = {
+        ...state,
+        [widgetId]: {
+          ...state[widgetId],
+          data: {
+            ...state[widgetId].data,
+            [field]: value,
+          },
+        },
+      };
+
+      return updatedWidgets;
+    }
   }
 
   return state;
@@ -48,4 +94,19 @@ export const widgetAdd = (widget: IWidgetProperties) => ({
 export const widgetRemove = (widgetId: string) => ({
   type: Actions.WidgetRemove,
   payload: widgetId,
+});
+
+export const widgetLocationUpdate = (widgetId: string, location: LocationTuple) => ({
+  type: Actions.WidgetLocationUpdate,
+  payload: { widgetId, location },
+});
+
+export const widgetDataUpdate = (widgetId: string, data: object) => ({
+  type: Actions.WidgetUpdateData,
+  payload: { widgetId, data },
+});
+
+export const widgetDataFieldUpdate = (widgetId: string, field: string, value: any) => ({
+  type: Actions.WidgetDataFieldUpdate,
+  payload: { widgetId, field, value },
 });